Move the archive helpers into the testing package
The lockfile check has to run as a datatest, which means the fingerprinting
code must be importable from an installed package rather than from scripts/.
The script keeps only the command line.
