+++ model/testing/tests/testing/unit_tests/data/v05/ser_data/ArchiveMetaData-icon_pydycore_rank0.json
