+++ model/testing/tests/testing/unit_tests/data/v06/NAMELIST_ICON_output_atm.json
